refactor: moved istype functions
